@LITSPHERE_e - book store
@LitSphere  This is our group name 
> Arjun Shaw(Main Coder)
> Ayan Das Adhikari (Web Page Designer)
> Arup Mondal (System Analyst)
> Arabinda Roy (Documentation)

This is to certify that the dissertation/project proposal entitled " E - BOOK STORE "
 is done by us, is an Authentic work carried out for the partial 
fulfilment of the requirements for the award of the certificate of COMPUTER SCIENCE AND 
TECHNOLOGY under the guidance of 'Mr. D . Chakraborty'. The matter embodied in this project 
work has not been submitted earlier for award of any certificate to the best of our 
knowledge and belief.
 
  GROUP MEMBERS_@litsphere                                                                       🙏Thank You🙏